Renamed config.ts to constants.ts
